requirements.txt
customtkinter==5.2.2
pandas==2.1.4
pymupdf==1.23.8
pillow==10.1.0